adicionando robo de input e salvando/lendo content em arquivo de json na raiz
