+++ .gitignore
@@ -62,3 +62,6 @@ typings/
 
 # credentials
 credentials/*.json
+
+#content
+content.json
