+++ robots/text.js
@@ -11,12 +11,17 @@ const nlu = new NaturalLanguageUnderstandingV1({
   url: 'https://gateway.watsonplatform.net/natural-language-understanding/api/'
 });
 
-async function robot(content) {
+const state = require('./state')
+
+async function robot() {
+  const content = state.load()
   await fetchContentFromWikipedia(content) //Pesquisa o conteúdo do wikipedia
   sanitizeContent(content) //Limpa o conteúdo retornado da busca
   breakContentIntoSentences(content) //Quebra o conteúdo retornado em frases
   limitMaximumSentences(content) //limita o numero de uso de frases no watson
   await fetchKeywordsOfAllSentences(content)
+
+  state.save(content)
 
   async function fetchContentFromWikipedia(content) {
     const algorithmiaAuthenticated = algorithmia(algorithmiaApiKey) //retorna uma instancia autenticada do algorithmia
